pc and mobile file, rearrange structure
