+++ lib/components/mobile/mobile_float_button.dart
@@ -1,7 +1,7 @@
 import 'package:animated_stack/animated_stack.dart';
 import 'package:flutter/material.dart';
 
-import '../../pages/mobile/sockets_page/sockets_page.dart';
+import '../../pages/mobile/sockets_page/mobile_sockets_page.dart';
 
 class MobileFloatButton extends StatefulWidget {
   final Widget mainPage;
