+++ lib/router/pc/pc_router.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:writing_writer/pages/pc/space_edit_page/space_edit_page.dart';
+import 'package:writing_writer/pages/pc/space_edit_page/pc_space_edit_page.dart';
 
 import '../../pages/common/pre_page/pre_page.dart';
-import '../../pages/pc/home_page/home_page.dart';
-import '../../pages/pc/sockets_page/sockets_page.dart';
+import '../../pages/pc/home_page/pc_home_page.dart';
+import '../../pages/pc/sockets_page/pc_sockets_page.dart';
 
 /// 需要引入跳转页面地址
